requirements.txt
pexpect
python-nmap
jsonpickle
